refactor: rename componets and add index.ts file to header
